User Repository Added for authentication and authorization purpose

+++ PostService/PostService/src/main/java/com/example/LinkedIn/PostService/repository/PostRepository.java
@@ -3,5 +3,5 @@ package com.example.LinkedIn.PostService.repository;
 import com.example.LinkedIn.PostService.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
 }
